Create PrivacyPolicy.md
What We Collect 
------------------
Borsalino's Revenge does not and will not collect personal information from the user.

Borsalino's Revenge stores the game data on the local device and this application does not transfer data off your device. 

What information we access or share
----------------------------------------
This application reads the device information to determine hardware and graphical capabilities. This data is kept local and not transferred anywhere.

This data can be deleted by removing the application from the user's local device.